Folder restructure committed for task 1
