+++ data/input_abp_1.txt
@@ -1,0 +1,3 @@
+00:00:10 5
+00:01:30 4
+00:20:00 3
