--- data_structures/message.cpp
@@ -1,25 +1,0 @@
-#include <math.h> 
-#include <assert.h>
-#include <iostream>
-#include <fstream>
-#include <string>
-
-#include "message.hpp"
-
-/***************************************************/
-/************* Output stream ************************/
-/***************************************************/
-
-ostream& operator<<(ostream& os, const Message_t& msg) {
-  os << msg.value;
-  return os;
-}
-
-/***************************************************/
-/************* Input stream ************************/
-/***************************************************/
-
-istream& operator>> (istream& is, Message_t& msg) {
-  is >> msg.value;
-  return is;
-}
