+++ include/data_structures/message.hpp
@@ -1,0 +1,31 @@
+#ifndef BOOST_SIMULATION_MESSAGE_HPP
+#define BOOST_SIMULATION_MESSAGE_HPP
+
+#include <assert.h>
+#include <iostream>
+#include <string>
+
+using namespace std;
+
+/*******************************************/
+/**************** Message_t ****************/
+/*******************************************/
+struct Message_t{
+
+  Message_t(){}
+  Message_t(float i_value)
+   :value(i_value){}
+
+  	float   value;
+
+  void clear() {
+   Message_t(); 
+  }
+};
+
+istream& operator>> (istream& is, Message_t& msg);
+
+ostream& operator<<(ostream& os, const Message_t& msg);
+
+
+#endif // BOOST_SIMULATION_MESSAGE_HPP
