+++ test/data/receiver_input_test.txt
@@ -1,0 +1,7 @@
+00:00:10 11
+00:00:30 20
+00:00:45 31
+00:00:52 31
+00:01:25 40
+00:01:35 40
+00:01:55 51
