+++ test/data/sender_input_test_control_In.txt
@@ -1,0 +1,4 @@
+00:00:00 -1
+00:00:05 0
+00:00:15 5
+00:02:50 3
