+++ test/data/subnet_input_test.txt
@@ -1,0 +1,20 @@
+00:00:10 11
+00:00:20 20
+00:00:30 31
+00:00:40 40
+00:00:50 51
+00:01:00 60
+00:01:10 71
+00:01:20 80
+00:01:30 91
+00:01:40 100
+00:01:50 111
+00:02:00 120
+00:02:10 131
+00:02:20 140
+00:02:30 151
+00:02:40 160
+00:02:50 171
+00:03:00 180
+00:03:10 191
+00:03:20 200
