--- test/sender/sender_input_test_ack_In.txt
@@ -1,7 +1,0 @@
-00:00:10 0
-00:00:30 1
-00:01:30 0
-00:01:55 1
-00:02:20 1
-00:02:45 0
-00:02:55 1
